api: split common tag functions to it own flow
In order to improve tag code, I've split the common functionality like
get tags, rename and delete into an own flow, from router until store
function.

It is important to note that none logic was changed in this commit, only
files were reordered.
