+++ api/services/device_tags_test.go
@@ -2,7 +2,6 @@ package services
 
 import (
 	"context"
-	"errors"
 	"testing"
 
 	storecache "github.com/shellhub-io/shellhub/api/cache"
@@ -85,7 +84,7 @@ func TestCreateTag(t *testing.T) {
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			tc.requiredMocks()
-			err := s.CreateTag(ctx, tc.uid, tc.deviceName)
+			err := s.CreateDeviceTag(ctx, tc.uid, tc.deviceName)
 			assert.Equal(t, tc.expected, err)
 		})
 	}
@@ -133,97 +132,7 @@ func TestRemoveTag(t *testing.T) {
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			tc.requiredMocks()
-			err := s.RemoveTag(ctx, tc.uid, tc.deviceName)
-			assert.Equal(t, tc.expected, err)
-		})
-	}
-
-	mock.AssertExpectations(t)
-}
-
-func TestRenameTag(t *testing.T) {
-	locator := &mocksGeoIp.Locator{}
-	mock := &mocks.Store{}
-	s := NewService(store.Store(mock), privateKey, publicKey, storecache.NewNullCache(), clientMock, locator)
-
-	ctx := context.TODO()
-
-	Err := errors.New("")
-
-	deviceWithTags := &models.Device{
-		UID:      "deviceWithTagsUID",
-		Name:     "deviceWithTagsName",
-		TenantID: "deviceWithTagsTenantID",
-		Tags:     []string{"device3", "device4", "device5"},
-	}
-
-	namespace := &models.Namespace{
-		Name:     "namespaceName",
-		Owner:    "owner",
-		TenantID: "namespaceTenantID",
-	}
-
-	cases := []struct {
-		name              string
-		tenantID          string
-		currentDeviceName string
-		newDeviceName     string
-		requiredMocks     func()
-		expected          error
-	}{
-		{
-			name:          invalidFormat,
-			tenantID:      "tenant",
-			requiredMocks: func() {},
-			expected:      ErrInvalidFormat,
-		},
-		{
-			name:              "Fail when device has no tags",
-			tenantID:          "namespaceTenantIDNoTag",
-			currentDeviceName: "device3",
-			newDeviceName:     "device1",
-			requiredMocks: func() {
-				mock.On("DeviceGetTags", ctx, "namespaceTenantIDNoTag").Return(nil, 0, Err)
-			},
-			expected: ErrNoTags,
-		},
-		{
-			name:              "Fail when device don't have a tag",
-			tenantID:          namespace.TenantID,
-			currentDeviceName: "device2",
-			newDeviceName:     "device1",
-			requiredMocks: func() {
-				mock.On("DeviceGetTags", ctx, namespace.TenantID).Return(deviceWithTags.Tags, len(deviceWithTags.Tags), nil).Once()
-			},
-			expected: ErrTagNameNotFound,
-		},
-		{
-			name:              "Fail when device already have a tag",
-			tenantID:          namespace.TenantID,
-			currentDeviceName: "device3",
-			newDeviceName:     "device5",
-			requiredMocks: func() {
-				mock.On("DeviceGetTags", ctx, namespace.TenantID).Return(deviceWithTags.Tags, len(deviceWithTags.Tags), nil).Once()
-			},
-			expected: ErrDuplicateTagName,
-		},
-		{
-			name:              "Successful rename a tag",
-			tenantID:          namespace.TenantID,
-			currentDeviceName: "device3",
-			newDeviceName:     "device1",
-			requiredMocks: func() {
-				mock.On("DeviceGetTags", ctx, namespace.TenantID).Return(deviceWithTags.Tags, len(deviceWithTags.Tags), nil).Once()
-				mock.On("DeviceRenameTag", ctx, namespace.TenantID, "device3", "device1").Return(nil).Once()
-			},
-			expected: nil,
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.name, func(t *testing.T) {
-			tc.requiredMocks()
-			err := s.RenameTag(ctx, tc.tenantID, tc.currentDeviceName, tc.newDeviceName)
+			err := s.RemoveDeviceTag(ctx, tc.uid, tc.deviceName)
 			assert.Equal(t, tc.expected, err)
 		})
 	}
@@ -292,119 +201,7 @@ func TestUpdateTag(t *testing.T) {
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			tc.requiredMocks()
-			err := s.UpdateTag(ctx, tc.uid, tc.tags)
-			assert.Equal(t, tc.expected, err)
-		})
-	}
-
-	mock.AssertExpectations(t)
-}
-
-func TestGetTags(t *testing.T) {
-	locator := &mocksGeoIp.Locator{}
-	mock := &mocks.Store{}
-	s := NewService(store.Store(mock), privateKey, publicKey, storecache.NewNullCache(), clientMock, locator)
-
-	ctx := context.TODO()
-
-	device := &models.Device{UID: "uid", Namespace: "namespace", TenantID: "tenant", Tags: []string{"device1", "device2"}}
-
-	namespace := &models.Namespace{Name: "namespace", TenantID: "tenant"}
-
-	type Expected struct {
-		Tags  []string
-		Count int
-		Error error
-	}
-
-	cases := []struct {
-		name          string
-		uid           models.UID
-		tenantID      string
-		requiredMocks func()
-		expected      Expected
-	}{
-		{
-			name:     "fail find the namespace",
-			tenantID: "not_found_tenant",
-			requiredMocks: func() {
-				mock.On("NamespaceGet", ctx, "not_found_tenant").Return(nil, ErrNotFound).Once()
-			},
-			expected: Expected{
-				Tags:  nil,
-				Count: 0,
-				Error: ErrNamespaceNotFound,
-			},
-		},
-		{
-			name:     "successful get tags",
-			tenantID: device.TenantID,
-			requiredMocks: func() {
-				mock.On("NamespaceGet", ctx, "tenant").Return(namespace, nil).Once()
-				mock.On("DeviceGetTags", ctx, "tenant").Return(device.Tags, len(device.Tags), nil).Once()
-			},
-			expected: Expected{
-				Tags:  device.Tags,
-				Count: len(device.Tags),
-				Error: nil,
-			},
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.name, func(t *testing.T) {
-			tc.requiredMocks()
-			returnedTags, count, err := s.GetTags(ctx, tc.tenantID)
-			assert.Equal(t, tc.expected, Expected{returnedTags, count, err})
-		})
-	}
-
-	mock.AssertExpectations(t)
-}
-
-func TestDeleteAllTags(t *testing.T) {
-	locator := &mocksGeoIp.Locator{}
-	mock := &mocks.Store{}
-	s := NewService(store.Store(mock), privateKey, publicKey, storecache.NewNullCache(), clientMock, locator)
-
-	ctx := context.TODO()
-
-	device := &models.Device{UID: "uid", Namespace: "namespace", TenantID: "tenant", Tags: []string{"device1", "device2"}}
-
-	namespace := &models.Namespace{Name: "namespace", TenantID: "tenant"}
-
-	cases := []struct {
-		name          string
-		deviceName    string
-		tenantID      string
-		requiredMocks func()
-		expected      error
-	}{
-		{
-			name:       "fail find the namespace",
-			deviceName: "device1",
-			tenantID:   "not_found_tenant",
-			requiredMocks: func() {
-				mock.On("NamespaceGet", ctx, "not_found_tenant").Return(nil, ErrNotFound).Once()
-			},
-			expected: ErrNamespaceNotFound,
-		},
-		{
-			name:       "successful get tags",
-			deviceName: "device1",
-			tenantID:   device.TenantID,
-			requiredMocks: func() {
-				mock.On("NamespaceGet", ctx, "tenant").Return(namespace, nil).Once()
-				mock.On("DeviceDeleteTags", ctx, "tenant", "device1").Return(nil).Once()
-			},
-			expected: nil,
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.name, func(t *testing.T) {
-			tc.requiredMocks()
-			err := s.DeleteTags(ctx, tc.tenantID, tc.deviceName)
+			err := s.UpdateDeviceTag(ctx, tc.uid, tc.tags)
 			assert.Equal(t, tc.expected, err)
 		})
 	}
